If I have to bugfix this God-forsaken api one more time, I am going to load a 2007 Ford Explorer XLT with 35 American pounds of Tannerite and drive it into where ever their Godless hovel of a headquarters are located.
Seriously, how the hell can they sleep at night with the knowledge of actually charging money from people to use this "service".  Everything about this api is a masterclass in what not to do to a user who wants to utilize your data for their own needs. Thank God we got this for free for "education" purposes, because frankly they deserve to be DDOS'd by a botnet of hundreds of thousands of IOT devices.

Anyway. Fixed it picking out incorrect values from the array they send back. It now searches based on the key in the json file, and if it doesn't find an entry, it sets it to, and displays, "No Data".