Fixed a bug (all instead of only half)
For whatever reason, this doesn't work with the "add to search" function. I don't get it, but it only seems to run the console logging commands inside the "add to search" method when the loop to remove non-us cities is active. Otherwise, it runs the code to add to the search, and returns the api, yet it says nothing is returned. I really don't get what is going on, and I'm sick of console logged debug messages only sometimes appearing when a completely separate loop, that is further ahead in the code, is active and not commented out.